Add another variable
Pretty sure my script is cursed, tried to make it work again and again, reloading and reloading. But nothing happened. Checked with regexr, and it worked there, but not on reddit where I was testing. Clicked on a permalink, and it worked, went back, and it worked... reloaded, and it worked... help, I don't know what I'm doing!